requirements.txt
chromadb
sentence-transformers
semantic-router
python-dotenv
ipykernel
ipywidgets